Refactor shell init code
